Move around ctor logic a bit
