+++ Src/VsInterfaces/VsInterfaces.csproj
