+++ Src/VsSpecific/VsSpecific/VsSpecific.projitems
@@ -9,6 +9,7 @@
     <Import_RootNamespace>VsSpecific</Import_RootNamespace>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="$(MSBuildThisFileDirectory)SharedServiceVersionFactory.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)SharedService.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)SharedService.Lazy.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)VsSpecificConstants.cs" />
